# docs/swagger.yaml
definitions:
  entities.SysApiServiceCreateReq:
    properties:
      description:
        type: string
      group:
        type: string
      is_use:
        type: integer
      method:
        type: integer
      path:
        type: string
    required:
    - description
    - group
    - method
    - path
    type: object
  entities.SysMenuServiceReq:
    properties:
      icon:
        type: string
      is_use:
        type: integer
      level:
        type: integer
      name:
        type: string
      pid:
        type: integer
      sort:
        type: integer
      unique_key:
        type: string
    required:
    - is_use
    - level
    - name
    - pid
    - unique_key
    type: object
  entities.UserServiceCreateReq:
    properties:
      nickname:
        type: string
      password:
        type: string
      phone:
        type: string
      status:
        type: integer
    required:
    - nickname
    - password
    - phone
    - status
    type: object
  entities.UserServiceTokenReq:
    properties:
      password:
        type: string
      phone:
        type: string
    type: object
  entities.UserServiceUpdateReq:
    properties:
      id:
        type: integer
      nickname:
        type: string
      phone:
        type: string
      status:
        type: integer
    type: object
info:
  contact: {}
paths:
  /sys_api:
    post:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: API基础信息
        in: body
        name: userInfo
        required: true
        schema:
          $ref: '#/definitions/entities.SysApiServiceCreateReq'
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 创建一条API记录
      tags:
      - API操作
  /sys_api/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 当前页
        in: query
        name: page
        required: true
        type: integer
      - description: 每页条目
        in: query
        name: pageSize
        required: true
        type: integer
      - description: 筛选条件
        in: query
        items:
          type: string
        name: searches
        type: array
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 获取api分页数据
      tags:
      - API操作
  /sys_menu:
    get:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 查询菜单列表
      tags:
      - 菜单操作
    post:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 菜单基础信息
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/entities.SysMenuServiceReq'
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 创建一条用户记录
      tags:
      - 菜单操作
  /user:
    post:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户基础信息
        in: body
        name: userInfo
        required: true
        schema:
          $ref: '#/definitions/entities.UserServiceCreateReq'
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 创建一条用户记录
      tags:
      - 用户操作
  /user/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户编号
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 查看用户记录
      tags:
      - 用户操作
    put:
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户编号
        in: path
        name: id
        required: true
        type: integer
      - description: 需要更新的用户信息
        in: body
        name: userInfo
        required: true
        schema:
          $ref: '#/definitions/entities.UserServiceUpdateReq'
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 更新用户信息
      tags:
      - 用户操作
  /user/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 手机号和密码
        in: body
        name: userInfo
        required: true
        schema:
          $ref: '#/definitions/entities.UserServiceTokenReq'
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 用户授权登录
      tags:
      - 用户操作
  /user/logout:
    post:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 退出会话
      tags:
      - 用户操作
swagger: "2.0"